cookies redefined: expired date and signed inserted validations with cookies if user logged in and is admin done at create, update and delete

+++ src/controllers/generic_controller.js
@@ -2,6 +2,7 @@ const paginationLinksService = require("../services/process_pagination_links");
 const validateDataService    = require("../services/process_validate_data");
 const dataUpdatesService     = require("../services/process_data_updates");
 const encryptionService      = require("../services/process_encryption");
+const cookieService          = require("../services/process_cookie");
 
 const backlogDB = require("../db/backlog_db");
 
@@ -9,15 +10,15 @@ const backlogDB = require("../db/backlog_db");
 
 function getAllItems(db, tableNameParam) {
    
-   return async function(req, res) {
-      
-      // To start building pagination
-      let { limit, offset } = req.query;
-      limit = limit && !isNaN(parseInt(limit)) ? parseInt(limit) : 5;
-      offset = offset && !isNaN(parseInt(offset)) ? parseInt(offset) : 0;
-      
+   return async function(req, res) {      
 
       try {
+
+         // To start building pagination
+         let { limit, offset } = req.query;
+         limit = limit && !isNaN(parseInt(limit)) ? parseInt(limit) : 5;
+         offset = offset && !isNaN(parseInt(offset)) ? parseInt(offset) : 0;
+
          const getTotalItems = await db.getTotalItems();
          const totalItems = getTotalItems[0].total_items;
 
@@ -53,10 +54,11 @@ function getAllItems(db, tableNameParam) {
 function getItemById(db, tableNameParam) {
    
    return async function(req, res) {
-      
-      const { id } = req.params;
 
       try {
+         
+         const { id } = req.params;
+
          const itemById = await db.getItemById(id);
          
          // To show item data of database
@@ -85,38 +87,51 @@ function getItemById(db, tableNameParam) {
 function addItem(db, tableNameParam) {
    
    return async function(req, res) {
+      
+      const userLoggedIn = cookieService.verifyCookie(req);
 
-      const bodyData = req.body;
-      let itemData;
+      if (userLoggedIn) {
+         if (userLoggedIn.userCategory === "admin") {
+      
+            try {
 
-      // To validate if any data is empty
-      await validateDataService.sanitiseBlankSpaces(itemData);
+               const bodyData = req.body;
+               let itemData;
 
-      // To validate if the data is about a user and if so to encrypt the password 
-      if (tableNameParam === "users") {
-         const { email, password, category } = bodyData;
-         const hashedPassword = await encryptionService.createHash(password);
-         itemData = { email, hashedPassword, category };
+               // To validate if any data is empty
+               await validateDataService.sanitiseBlankSpaces(bodyData);
 
+               // To validate if the data is about a user and if so to encrypt the password 
+               if (tableNameParam === "users") {
+                  const { email, password, category } = bodyData;
+                  const hashedPassword = await encryptionService.createHash(password);
+                  itemData = { email, hashedPassword, category };
+
+               } else {
+                  itemData = bodyData;
+               };
+            
+               const newItem = await db.addItem(itemData);
+
+               // To show item data inserted in database
+               const lastItemInserted = newItem.insertId;
+               const item = await db.getItemById(lastItemInserted);
+               res.json(item[0]);
+
+               // To log into backlog if any insert have been done
+               const action = "insert";
+               const tableName = tableNameParam;
+               await backlogDB.logChangesToBacklog(item[0], item[0].id, action, tableName);
+
+            } catch(error) {
+               res.status(500).send(error.message);
+            };
+
+         } else {
+            res.status(401).send("WARNING: User not authorized!");
+         };
       } else {
-         itemData = bodyData;
-      };
-   
-      try {
-         const newItem = await db.addItem(itemData);
-
-         // To show item data inserted in database
-         const lastItemInserted = newItem.insertId;
-         const item = await db.getItemById(lastItemInserted);
-         res.json(item[0]);
-
-         // To log into backlog if any insert have been done
-         const action = "insert";
-         const tableName = tableNameParam;
-         await backlogDB.logChangesToBacklog(item[0], item[0].id, action, tableName);
-
-      } catch(error) {
-         res.status(500).send(error.message);
+         res.status(401).send("WARNING: Please, login!");
       };
    };
 };
@@ -127,35 +142,47 @@ function updateItem(db, tableNameParam) {
    
    return async function(req, res) {
 
-      const { id } = req.params;
-      const itemData = req.body;
+      const userLoggedIn = cookieService.verifyCookie(req);
 
-      try {
-         const oldItemData = await db.getItemById(id);
+      if (userLoggedIn) {
+         if (userLoggedIn.userCategory === "admin") {
          
-         const itemUpdated = await db.updateItem(id, itemData);
+            try {
+               const { id } = req.params;
+               const itemData = req.body;
 
-         // To show item data updated in database
-         const item = itemUpdated.affectedRows;
-         
-         if (item === 1) {
-            const newItemData = await db.getItemById(id);
-            res.json(newItemData[0]);
-            
-            // To log into backlog if any update have been done
-            const oldData = oldItemData[0];
-            const newData = newItemData[0];
-            const itemDataUpdated = dataUpdatesService.processDataUpdates(oldData, newData);
-            const action = "update";
-            const tableName = tableNameParam;
-            await backlogDB.logChangesToBacklog(itemDataUpdated, id, action, tableName);
-
+               const oldItemData = await db.getItemById(id);
+               
+               const itemUpdated = await db.updateItem(id, itemData);
+      
+               // To show item data updated in database
+               const item = itemUpdated.affectedRows;
+               
+               if (item === 1) {
+                  const newItemData = await db.getItemById(id);
+                  res.json(newItemData[0]);
+                  
+                  // To log into backlog if any update have been done
+                  const oldData = oldItemData[0];
+                  const newData = newItemData[0];
+                  const itemDataUpdated = dataUpdatesService.processDataUpdates(oldData, newData);
+                  const action = "update";
+                  const tableName = tableNameParam;
+                  await backlogDB.logChangesToBacklog(itemDataUpdated, id, action, tableName);
+      
+               } else {
+                  res.status(404).send("WARNING: Item not found!");
+               }; 
+      
+            } catch(error) {
+               res.status(500).send(error.message);
+            };
+   
          } else {
-            res.status(404).send("WARNING: Item not found!");
-         }; 
-
-      } catch(error) {
-         res.status(500).send(error.message);
+            res.status(401).send("WARNING: User not authorized!");
+         };
+      } else {
+         res.status(401).send("WARNING: Please, login!");
       };
    };
 };
@@ -166,31 +193,44 @@ function deleteItem(db, tableNameParam) {
    
    return async function(req, res) {
 
-      const { id } = req.params;
+      const userLoggedIn = cookieService.verifyCookie(req);
 
-      try {
-         const oldItemData = await db.getItemById(id);
-         const itemDeleted = await db.deleteItem(id);
-         
-         // To show item data deleted from database
-         const item = itemDeleted.affectedRows;
+      if (userLoggedIn) {
+         if (userLoggedIn.userCategory === "admin") {
+      
+            try {
 
-         if (item === 1) {
-            await db.getItemById(id);
-            res.send("WARNING: Item deleted!");
+               const { id } = req.params;
 
-            // To log into backlog if any delete have been done
-            const oldData = oldItemData[0];
-            const action = "delete";
-            const tableName = tableNameParam;
-            await backlogDB.logChangesToBacklog(oldData, id, action, tableName);
+               const oldItemData = await db.getItemById(id);
+               const itemDeleted = await db.deleteItem(id);
+               
+               // To show item data deleted from database
+               const item = itemDeleted.affectedRows;
+
+               if (item === 1) {
+                  await db.getItemById(id);
+                  res.status(200).send("SUCCESS: Item deleted!");
+
+                  // To log into backlog if any delete have been done
+                  const oldData = oldItemData[0];
+                  const action = "delete";
+                  const tableName = tableNameParam;
+                  await backlogDB.logChangesToBacklog(oldData, id, action, tableName);
+
+               } else {
+                  res.status(404).send("WARNING: Item not found!");
+               }; 
+
+            } catch(error) {
+               res.status(500).send(error.message);
+            };
 
          } else {
-            res.status(404).send("WARNING: Item not found!");
-         }; 
-
-      } catch(error) {
-         res.status(500).send(error.message);
+            res.status(401).send("WARNING: User not authorized!");
+         };
+      } else {
+         res.status(401).send("WARNING: Please, login!");
       };
    };
 };
